✨ feat: added jwt verify

+++ server/src/routes/memories.ts
@@ -1,13 +1,25 @@
-import { FastifyInstance } from "fastify";
 import { z } from "zod";
+import { FastifyInstance } from "fastify";
 import { prisma } from "../lib/prisma";
 
 // Neste arquivo ficará todas os métodos HTTPS que forem relacionados as memories
 
 export async function memoriesRoutes(app: FastifyInstance) {
+  // antes de executar as rotas, vai verificar se o usuário está logado
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify();
+  });
+
   // lista todos os campos ordenados pela data de criação em ordem decrescente da tabela "memory"
-  app.get("/memories", async () => {
+  app.get("/memories", async (request) => {
+    // verifica se quando o usuário acessar a rota, ele está com o token jwt. Caso contrário,
+    // bloqueia para que não consiga acessar a rota. Só pode acessar se tiver logado
+    await request.jwtVerify();
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -24,7 +36,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
   });
 
   // lista uma memória em específico
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, reply) => {
     // coleta um objeto ( que será o id ) e verifica se ele é uma string e um UUID
     const paramsSchema = z.object({ id: z.string().uuid() });
 
@@ -37,6 +49,12 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     });
+
+    // se a memória não for pública e o id do usuário for diferente
+    // do usuário da memória, retorna um erro
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
     return memory;
   });
@@ -63,7 +81,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "436270c3-ff5c-4e18-b6d6-3b32869a3a22",
+        userId: request.user.sub,
       },
     });
 
@@ -71,7 +89,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
   });
 
   // atualiza uma memória
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({ id: z.string().uuid() });
     const { id } = paramsSchema.parse(request.params);
 
@@ -82,8 +100,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    // se o id do usuário for diferente do usuário da memória, retorna um erro
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
     // atualiza o content, coverUrl, isPublic onde o id passado na requisição for o mesmo da tabela memory
-    const memory = await prisma.memory.update({
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -98,10 +127,21 @@ export async function memoriesRoutes(app: FastifyInstance) {
   });
 
   // deleta uma memória
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({ id: z.string().uuid() });
 
     const { id } = paramsSchema.parse(request.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    // se o id do usuário for diferente do usuário da memória, retorna um erro
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
     await prisma.memory.delete({
       where: {
